+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Known Issues:
 - None
 
+## v0.2.2 - Nate Haddad, 4/6/2025
+- Add Agentic RAG retriever tool to `SmolAgentsAnalyst`
+- Update `SmolAgentsAnalyst` to use memory, set `max_steps`
+
 ## v0.2.1 - Nate Haddad, 4/5/2025
 - Add new `tool` for getting image caption using OpenAI